Si47XX Arduino Library

+++ examples/SI47XX_KITS/THIAGO_LIMA/SI473X_TOUCH_THIAGO_BETA_V3/SI473X_TOUCH_THIAGO_BETA_V3.ino
@@ -187,8 +187,8 @@ uint8_t MinAGCgain         =  1;
 
 
 
-uint8_t currentVOL         =  0;
-uint8_t previousVOL        =  0;
+int currentVOL         =  0;
+int previousVOL        =  0;
 uint8_t currentVOLStep     =  1;
 uint8_t MaxVOL             = 63;
 uint8_t MinVOL             =  0;
